% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_sample_sim.R
\name{bulk_sample_tumor}
\alias{bulk_sample_tumor}
\title{Bulk sample tumor}
\usage{
bulk_sample_tumor(
  alive_cells,
  n_samples = 25,
  punch_radius_n_cells = 2,
  acceptable_overlap = 1,
  max_sampling_attempts = 100,
  max_shifting_attempts = 10,
  max_punch_size_deviation = 5
)
}
\arguments{
\item{alive_cells}{data.frame of alive cells in simulated tumor}

\item{n_samples}{integer number of punches to sample}

\item{punch_radius_n_cells}{integer radius of punches in number of cells}

\item{acceptable_overlap}{number of cells punches can overlap by}

\item{max_sampling_attempts}{number of sampling attempts before aborted}

\item{max_shifting_attempts}{number of punch shifts attempted before punch is re-drawn}

\item{max_punch_size_deviation}{accepted devation from expected punch size, due to punch overlap or edge}
}
\value{
list of punched cells
}
\description{
Select non-overlapping punch biopsies to sample from simulated tumor
}
