% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{calc_dist_from_center}
\alias{calc_dist_from_center}
\title{Calc distance from center}
\usage{
calc_dist_from_center(tree)
}
\arguments{
\item{tree}{treedata}
}
\value{
treedata
}
\description{
Calculate distance of cell from tumor center.
}
