% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{calc_freq_spectrums}
\alias{calc_freq_spectrums}
\title{Calculate frequency spectrums}
\usage{
calc_freq_spectrums(muts, death_rate)
}
\arguments{
\item{muts}{data.frame}

\item{death_rate}{numeric}
}
\value{
data.frame
}
\description{
Calculate frequency spectrums from all simulations with given death rate
}
