% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{calc_lbi}
\alias{calc_lbi}
\title{Calculate LBI}
\usage{
calc_lbi(tree, muts, tau = 0)
}
\arguments{
\item{tree}{treedata}

\item{muts}{data.frame}

\item{tau}{numeric}
}
\value{
treedata
}
\description{
Calculate LBI statistic on tree nodes.
}
