% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{calc_pairwise_genetic_distance}
\alias{calc_pairwise_genetic_distance}
\title{Calculate pairwise genetic distance}
\usage{
calc_pairwise_genetic_distance(child, ancestor, all_muts)
}
\arguments{
\item{child}{integer}

\item{ancestor}{integer}

\item{all_muts}{data.frame}
}
\value{
numeric
}
\description{
Calculate molecular branch lengths when generating tree given child and ancestor indices.
}
