% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{calc_segment_edge_time}
\alias{calc_segment_edge_time}
\title{Calc segment edge time}
\usage{
calc_segment_edge_time(node, tree, sim_cells)
}
\arguments{
\item{node}{integer}

\item{tree}{treedata}

\item{sim_cells}{data.frame}
}
\value{
numeric
}
\description{
Calculate time spend on edge of branch
}
