% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{calc_spatial_distance_given_leaves}
\alias{calc_spatial_distance_given_leaves}
\title{Calc distance from center}
\usage{
calc_spatial_distance_given_leaves(tree, sp1, sp2)
}
\arguments{
\item{tree}{treedata}

\item{sp1}{integer}

\item{sp2}{integer}
}
\value{
numeric
}
\description{
Calculate distance of cell from tumor center.
}
