% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mcmc.R
\name{calc_state_dependent_growthRateDiff_means}
\alias{calc_state_dependent_growthRateDiff_means}
\title{Calculate state-dependent growth rate difference}
\usage{
calc_state_dependent_growthRateDiff_means(mcmc_obj, dr, fixed_dr = NULL)
}
\arguments{
\item{mcmc_obj}{mcmc logs which include birthRateCanonical.1 and birthRateCanonical.0}

\item{dr}{true simulated death rate, if NULL will look for death rate in deathRateCanonical.1 and deathRateCanonical.0}

\item{fixed_dr}{parameter given to BEAST inference, if NULL will look for death rate in deathRateCanonical.1 and deathRateCanonical.0}
}
\value{
data.frame with mean and HPD intervals
}
\description{
Calculate mean and HPD intervals for difference between edge and center growth rates
}
