% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{check_if_cell_most_extreme}
\alias{check_if_cell_most_extreme}
\title{Check if cell is most extreme.}
\usage{
check_if_cell_most_extreme(x_loc, y_loc, alive_cells)
}
\arguments{
\item{x_loc}{numeric}

\item{y_loc}{numeric}

\item{alive_cells}{data.frame}
}
\value{
integer
}
\description{
Check if cell is at an extreme x or y location in the tumor to determine edge.
}
