% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{cluster_cells}
\alias{cluster_cells}
\title{Cluster cells}
\usage{
cluster_cells(sampled_cells, k = 10)
}
\arguments{
\item{sampled_cells}{data.frame}

\item{k}{data.frame}
}
\value{
numeric
}
\description{
Cluster cells from genetic data
}
