% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{collect_all_ancestors}
\alias{collect_all_ancestors}
\title{Collect all ancestors}
\usage{
collect_all_ancestors(index, all_cells)
}
\arguments{
\item{index}{integer}

\item{all_cells}{data.frame}
}
\value{
vector
}
\description{
get all ancestors traversing up tree of indexed node.
}
