% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{compare_muts}
\alias{compare_muts}
\title{Compare mutations}
\usage{
compare_muts(row_muts, all_muts)
}
\arguments{
\item{row_muts}{vector}

\item{all_muts}{data.frame}
}
\value{
numeric
}
\description{
Compare all mutations with row.
}
