% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{convert_all_cells_to_tree_fast}
\alias{convert_all_cells_to_tree_fast}
\title{Make full tree from simulated cells fast}
\usage{
convert_all_cells_to_tree_fast(
  all_cells,
  add_all_states = FALSE,
  sampled_cells_indices = NULL,
  branch_unit = "time",
  cell_locations_df_file = NULL
)
}
\arguments{
\item{all_cells}{data.frame of all simulated cells in tumor}

\item{add_all_states}{logical. Default FALSE. If included edge/center states, which takes more time.}

\item{sampled_cells_indices}{Vector of indices of sampled cells. If NULL, all cells will be included in tree.}

\item{branch_unit}{string. Default "time". Other options are "genetic" and "none".}

\item{cell_locations_df_file}{File to find locations through time for pushing simulation.}
}
\value{
numeric
}
\description{
Make tree from simulated cells including spatial constraints but faster by converted to nwk string first.
Use prune_simulated tree to convert to subsampled tree.
}
