% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{convert_nodes_to_string}
\alias{convert_nodes_to_string}
\title{Convert nodes to string}
\usage{
convert_nodes_to_string(
  orphan_cells,
  all_cells,
  branch_unit = c("time", "molecular", "generations", "none")
)
}
\arguments{
\item{orphan_cells}{vector of sampled leaves}

\item{all_cells}{data.frame of all cells in simulation}

\item{branch_unit}{Unit to calcualte branch length. Options: "time", "molecular", "generations", "none".}
}
\value{
list
}
\description{
Convert sampled simulated cells into newick formatted string and node list.
}
