% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{convert_nwk_to_treedata}
\alias{convert_nwk_to_treedata}
\title{Convert newick to treedata}
\usage{
convert_nwk_to_treedata(
  nwk_list,
  all_cells,
  sampled_cells,
  include_edge_states = FALSE
)
}
\arguments{
\item{nwk_list}{list of newick formated tree string and vector of indices in tree. Output of convert_nodes_to_string.}

\item{all_cells}{data.frame of all cells in the simulation}

\item{sampled_cells}{data.frame of sampled alive cells}

\item{include_edge_states}{logical to specify if edge states will be included in treedata. Default FALSE}
}
\value{
treedata
}
\description{
Convert tree in newick formatted string to ggtree object with location data.
For easy plotting and use with R phylogenetic tools
}
