% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_to_nexus.R
\name{convert_sim_to_nex}
\alias{convert_sim_to_nex}
\title{Convert JC sim to nexus}
\usage{
convert_sim_to_nex(
  sim_cells_file,
  nexus_file,
  sampled_cells_file = NULL,
  n_sampled = 100,
  new_sampled_cells_file = NULL,
  type = c("tree", "topology", "alignment", "states"),
  diversified_sampling = FALSE,
  edge_weight = 0.5,
  fixed_edge_num = NULL
)
}
\arguments{
\item{sim_cells_file}{string}

\item{nexus_file}{string}

\item{sampled_cells_file}{string}

\item{n_sampled}{integer}

\item{new_sampled_cells_file}{string}

\item{type}{string}

\item{diversified_sampling}{FALSE, if TRUE will select from categories based on weight}

\item{edge_weight}{weight for random diversified sampling}

\item{fixed_edge_num}{number of edge states to sample}
}
\value{
numeric
}
\description{
Write simulated tumor to nexus file.
}
