% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{convert_sim_to_tree}
\alias{convert_sim_to_tree}
\title{Convert simulation to tree}
\usage{
convert_sim_to_tree(
  sampled_cells,
  all_cells,
  branch_unit = "time",
  include_edge_states = FALSE
)
}
\arguments{
\item{sampled_cells}{data.frame}

\item{all_cells}{data.frame}

\item{branch_unit}{string}

\item{include_edge_states}{logical}
}
\value{
tree
}
\description{
Convert data.frame of simulated cell to g to ggtree object with location data.
}
