% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{count_spatial_pops_at_t}
\alias{count_spatial_pops_at_t}
\title{Count spatial populations at t}
\usage{
count_spatial_pops_at_t(all_cells, time)
}
\arguments{
\item{all_cells}{data.frame}

\item{time}{numeric}
}
\value{
data.frame
}
\description{
Count population of spatial comparments in tumor at time t.
}
