% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_tree.R
\name{create_node_conversion_chart}
\alias{create_node_conversion_chart}
\title{Create node conversion chart}
\usage{
create_node_conversion_chart(sim_tree, est_tree)
}
\arguments{
\item{sim_tree}{treedata}

\item{est_tree}{treedata}
}
\value{
data.frame
}
\description{
Make data frame that matches simulated nodes to 
nodes in estimated tree
}
