% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{define_mut_presence_absence}
\alias{define_mut_presence_absence}
\title{Define mutations presence-absence}
\usage{
define_mut_presence_absence(sampled_cells)
}
\arguments{
\item{sampled_cells}{data.frame}
}
\value{
data.frame
}
\description{
Create presence-absence data.frame of samples vs sampled mutations.
}
