% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{diversified_sampling}
\alias{diversified_sampling}
\title{Diversified sampling}
\usage{
diversified_sampling(alive_cells, n_sampled_cells)
}
\arguments{
\item{alive_cells}{data.frame Alive cells in tumor}
}
\value{
data.frame
}
\description{
Samples alive cells from simulated tumor with maximized distance between points
}
