% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mcmc.R
\name{extract_death_rate}
\alias{extract_death_rate}
\title{Extract death rate}
\usage{
extract_death_rate(filename)
}
\arguments{
\item{filename}{}
}
\value{
dr numeric
}
\description{
Extract simulated death rate from name of file
}
