% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_to_nexus.R
\name{extract_mutations}
\alias{extract_mutations}
\title{Extract mutations}
\usage{
extract_mutations(mut_row)
}
\arguments{
\item{mut_row}{integer}
}
\value{
vector
}
\description{
Extract binary mutations from cells data frame
}
