% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_sample_sim.R
\name{extract_punch_allele_freqs}
\alias{extract_punch_allele_freqs}
\title{Extract punch allele freqs}
\usage{
extract_punch_allele_freqs(id, mut_presence_absence_df, sampled_cells)
}
\arguments{
\item{id}{integer id of punch}

\item{mut_presence_absence_df}{data.frame of binary variant presense absence status for each variant}

\item{sampled_cells}{data.frame of sampled cells}
}
\description{
Find true allele frequencies of sampled frequencies in punch
}
