% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_to_nexus.R
\name{extract_sequences}
\alias{extract_sequences}
\title{Extract sequences}
\usage{
extract_sequences(seq, sampled_muts = NULL)
}
\arguments{
\item{seq}{string}

\item{sampled_muts}{vector, if NULL then sequence will not be filtere to observed variation}

\item{type}{string}
}
\value{
string
}
\description{
Extract sequences from seq column
}
