% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{filter_alive_cells}
\alias{filter_alive_cells}
\title{Filter alive cells}
\usage{
filter_alive_cells(cells_df)
}
\arguments{
\item{cells_df}{data.frame}
}
\value{
data.frame
}
\description{
Filter cells to those alive at the end of the simulation.
}
