% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{find_all_cell_state_at_death}
\alias{find_all_cell_state_at_death}
\title{Get all cell states at death}
\usage{
find_all_cell_state_at_death(all_cells, cell_locations_df_file = NULL)
}
\arguments{
\item{all_cells}{data.frame}

\item{cell_locations_df_file}{CSV file of cell locations. Default: null}
}
\value{
number of cell divisions
}
\description{
Get states of cells as death by time slice, this method is faster than going cell by cell.
}
