% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{find_cell_state_at_death}
\alias{find_cell_state_at_death}
\title{Find cell state at death}
\usage{
find_cell_state_at_death(index, all_cells, cell_locations_df_file = NULL)
}
\arguments{
\item{index}{integer of cell index}

\item{all_cells}{data.frame of all simulated cells}

\item{cell_locations_df_file}{CSV file for locations of cells through time. If NULL, which is default, assume current locations have not changed.}
}
\value{
integer
}
\description{
Find binary edge/center spatial state at time of cell division or death
}
