% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{find_MRCA}
\alias{find_MRCA}
\title{Find MRCA}
\usage{
find_MRCA(index_1, index_2, all_cells)
}
\arguments{
\item{index_1}{integer}

\item{index_2}{integer}

\item{all_cells}{data.frame}
}
\value{
integer
}
\description{
Find MRCA between two cells given indexes.
}
