% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_all_states_cell}
\alias{get_all_states_cell}
\title{Get all cell states}
\usage{
get_all_states_cell(index, all_cells)
}
\arguments{
\item{index}{integer}

\item{all_cells}{data.frame}
}
\value{
data.frame
}
\description{
Get spatial state of cell over lifespan.
}
