% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_cell_pops_over_time}
\alias{get_cell_pops_over_time}
\title{Get spatial populations at over time}
\usage{
get_cell_pops_over_time(all_cells)
}
\arguments{
\item{all_cells}{data.frame}
}
\value{
data.frame
}
\description{
Create time points and get populations over time
}
