% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_cell_time_on_est_edge}
\alias{get_cell_time_on_est_edge}
\title{Get cell time on est edge}
\usage{
get_cell_time_on_est_edge(cell_index, sim_cells, timestep = 1/24)
}
\arguments{
\item{cell_index}{integer}

\item{sim_cells}{data.frame}

\item{timestep}{numeric}
}
\value{
data.frame
}
\description{
Get time cells spends on esimated edge
}
