% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get color palette}
\usage{
get_color_palette(
  names = c("edge", "center", "boundary_driven", "unrestricted")
)
}
\arguments{
\item{names}{vector of names to include for colors. For example "edge", "center"}
}
\value{
vector
}
\description{
Get color palette to use to plot edge versus center, boundary-driven versus unrestricted growth.
}
