% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_compartment_growth_rates_at_t}
\alias{get_compartment_growth_rates_at_t}
\title{Get spatial compartment growth rates}
\usage{
get_compartment_growth_rates_at_t(all_cells, time, time_interval = 1/24)
}
\arguments{
\item{all_cells}{data.frame}

\item{time}{numeric}

\item{time_interval}{numeric}
}
\value{
numeric
}
\description{
Extract instantaneous edge and center growth rates over time
}
