% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_edge_dist_versus_divisions_at_t}
\alias{get_edge_dist_versus_divisions_at_t}
\title{Edge distances versus cell divisions over time}
\usage{
get_edge_dist_versus_divisions_at_t(all_cells, time, time_interval = 1/24)
}
\arguments{
\item{all_cells}{data.frame}

\item{time}{numeric}

\item{time_interval}{numeric}
}
\value{
numeric
}
\description{
Extract instantaneous fates of cells over simulation time and record distance from current tumor edge
}
