% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_edge_time_cell}
\alias{get_edge_time_cell}
\title{Get cell edge time}
\usage{
get_edge_time_cell(index, all_cells)
}
\arguments{
\item{index}{integer}

\item{all_cells}{data.frame}
}
\value{
data.frame
}
\description{
Given index, get how much time a cell spent on the edge of the tumor in its lifetime and
how long it lived.
}
