% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_free_cells_at_t}
\alias{get_free_cells_at_t}
\title{Get free cells at time t}
\usage{
get_free_cells_at_t(index, time, all_cells)
}
\arguments{
\item{index}{integer}

\item{time}{numeric}

\item{all_cells}{data.frame}
}
\value{
integer
}
\description{
Given index, get the number of free surrounding lattice spots of cells at time t.
}
