% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_lineage_segment}
\alias{get_lineage_segment}
\title{Get lineage segment}
\usage{
get_lineage_segment(node, tree, sim_cells)
}
\arguments{
\item{node}{integer}

\item{tree}{treedata}
}
\value{
vector
}
\description{
Collect all cells of branch
}
