% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_lineage_states}
\alias{get_lineage_states}
\title{Get all cell states of lineage}
\usage{
get_lineage_states(leaf, ancestor = 1, all_cells)
}
\arguments{
\item{leaf}{integer}

\item{ancestor}{integer}

\item{all_cells}{data.frame}
}
\value{
data.frame
}
\description{
Get spatial states of full lineage over history.
}
