% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_lineage_time_on_edge}
\alias{get_lineage_time_on_edge}
\title{Get lineage time on edge}
\usage{
get_lineage_time_on_edge(leaf, ancestor = 1, all_cells)
}
\arguments{
\item{leaf}{integer}

\item{ancestor}{integer}

\item{all_cells}{data.frame}
}
\value{
data.frame
}
\description{
Given leaf, calculate how long the total lineage from root to leaf spent on the edge.
how long it lived.
}
