% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_n_adjacent_cells}
\alias{get_n_adjacent_cells}
\title{Get n adjacent cells}
\usage{
get_n_adjacent_cells(x_loc, y_loc, alive_cells)
}
\arguments{
\item{x_loc}{numeric}

\item{y_loc}{numeric}

\item{all_cells}{data.frame}
}
\value{
data.frame
}
\description{
Given x-y location and current alive cells, find how many free adjacent lattice spots there are.
}
