% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{get_n_divisions}
\alias{get_n_divisions}
\title{Get n divisions}
\usage{
get_n_divisions(index, all_cells)
}
\arguments{
\item{index}{numerical index of extant cell}

\item{all_cells}{data.frame}
}
\value{
number of cell divisions
}
\description{
Get number of divisions along lineage of extant cells
}
