% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_sample_sim.R
\name{get_punched_cells}
\alias{get_punched_cells}
\title{Get punched cells}
\usage{
get_punched_cells(sampled_cell_index, alive_cells, punch_radius_n_cells)
}
\arguments{
\item{sampled_cell_index}{integer sampled cell that is center of punch}

\item{alive_cells}{data.frame all alive cells in simulated tumor}

\item{punch_radius_n_cells}{integer number of cells from index cells to punch, value0 will return single cels}
}
\value{
data.frame
}
\description{
Collect punch in radius surrounding sampled cell
}
