% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{get_row_muts}
\alias{get_row_muts}
\title{Get row mutations}
\usage{
get_row_muts(mutations_row)
}
\arguments{
\item{mutations_row}{integer}
}
\value{
vector
}
\description{
Extract vector of 0-1 mutations from string.
}
