% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{get_row_muts_by_i}
\alias{get_row_muts_by_i}
\title{Get row mutations by i}
\usage{
get_row_muts_by_i(i, mutations_column)
}
\arguments{
\item{i}{integer}

\item{mutations_column}{integer}
}
\value{
vector
}
\description{
Alternate function to extract mutations by column.
}
