% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{isolate_muts}
\alias{isolate_muts}
\title{Isolate sampled mutations}
\usage{
isolate_muts(mutations_column)
}
\arguments{
\item{mutations_column}{integer}
}
\value{
vector
}
\description{
Generate presence-absence matrix based on sampled mutations.
}
