% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{label_node_types}
\alias{label_node_types}
\title{Label node types}
\usage{
label_node_types(tree)
}
\arguments{
\item{tree}{treedata}
}
\value{
treedata
}
\description{
Label leaves and nodes to easily separate
}
