% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_tree.R
\name{make_calbriation_df}
\alias{make_calbriation_df}
\title{Make calibration dataframe}
\usage{
make_calbriation_df(sim_tree, est_tree, type = c("revbayes", "bdmm"))
}
\arguments{
\item{sim_tree}{treedata}

\item{est_tree}{treedata}

\item{type}{string}
}
\value{
data.frame
}
\description{
Compare the probability of edge state with 
the true simulated states for internal nodes
of the esimated tree.
}
