% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{make_tree_from_simulated_cells}
\alias{make_tree_from_simulated_cells}
\title{Make tree from simulated cells}
\usage{
make_tree_from_simulated_cells(
  sampled_cells,
  sim_cells,
  add_cell_to_label = TRUE,
  branch_unit = "time",
  include_edge_states = TRUE
)
}
\arguments{
\item{sampled_cells}{data.frame}

\item{sim_cells}{data.frame}

\item{add_cell_to_label}{logical. Default TRUE}

\item{branch_unit}{string. Default "time". Other options are "molecular", "none".}

\item{include_edge_states}{string}
}
\value{
numeric
}
\description{
Make tree from simulated cells including spatial constraints.
}
