% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{mark_boundary}
\alias{mark_boundary}
\title{Mark boundary}
\usage{
mark_boundary(cells_to_mark, alive_cells)
}
\arguments{
\item{cells_to_mark}{data.frame}

\item{alive_cells}{data.frame}
}
\value{
data.frame
}
\description{
Mark if set of cells are on boundary given alive cells.
}
