% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_tree.R
\name{match_sim_est_node}
\alias{match_sim_est_node}
\title{Match simlated and estimated node}
\usage{
match_sim_est_node(node, sim_tree, est_tree)
}
\arguments{
\item{node}{integer}

\item{sim_tree}{treedata}

\item{est_tree}{treedata}
}
\value{
integer
}
\description{
Match node in simulated tree to node in estimated tree
based on set of child leaves descended from node.
}
