% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{MCC_colour_plot}
\alias{MCC_colour_plot}
\title{Plot MCBD Tree}
\usage{
MCC_colour_plot(
  treefile,
  plotfile = NULL,
  scalel = NULL,
  scalem = NULL,
  colour_gradient = c("red", "yellow", "green")
)
}
\arguments{
\item{treefile}{file containing the MCC tree}

\item{plotfile}{output file. If NULL, plots are sent to the R console.}

\item{scalel}{min and max values of lambda to use for the colour scale. If NULL, calculated from the tree values.}

\item{scalem}{min and max values of mu to use for the colour scale. If NULL, calculated from the tree values.}

\item{colour_gradient}{colours used on the scale, in order from low to high values.}
}
\description{
Plot an MCC tree with median lambda / mu values on a coloured scale
}
