% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{normalize_locs}
\alias{normalize_locs}
\title{Normalize locations}
\usage{
normalize_locs(cells)
}
\arguments{
\item{cells}{data.frame}
}
\value{
data.frame
}
\description{
Normalize x-y coordinates in of simulated cells.
}
