% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{plot_space_phylo}
\alias{plot_space_phylo}
\title{Plot spatial phylogeny}
\usage{
plot_space_phylo(tree, spatial_file = "spatial.png")
}
\arguments{
\item{tree}{treedata}

\item{spatial_file}{string}
}
\description{
Plot tree on x-y coordinates.
}
