% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{process_and_sample_cells}
\alias{process_and_sample_cells}
\title{Process and sample cells}
\usage{
process_and_sample_cells(cells, n)
}
\arguments{
\item{n}{integer}

\item{alive_cells}{data.frame}
}
\value{
data.frame
}
\description{
Wrapper function to filter and sample alive cells at the end of the simulation.
}
