% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{prune_simulated_tree}
\alias{prune_simulated_tree}
\title{Prune simulated tree}
\usage{
prune_simulated_tree(
  tree,
  sampled_cells_indices,
  add_all_states = FALSE,
  all_cells = NULL,
  cell_locations_df_file = NULL,
  branch_unit = "time"
)
}
\arguments{
\item{tree}{treeio S4 tree of simulated tumor}

\item{sampled_cells_indices}{Vector of indices of sampled cells. If NULL, all cells will be included in tree.}

\item{add_all_states}{logical. Default FALSE. If included edge/center states, which takes more time.}

\item{all_cells}{data.frame of all simulated cells. Default NULL but necessary to get edge/center states.}

\item{cell_locations_df_file}{File to find locations through time for pushing simulation.}
}
\value{
numeric
}
\description{
Prune full simulated tree to subsampled tips
}
