% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_bisse.R
\name{read_log_file}
\alias{read_log_file}
\title{Read log files}
\usage{
read_log_file(file, percent.burnin = 0.1)
}
\arguments{
\item{file}{log file}

\item{percent.burnin}{fraction of mcmc burnin}
}
\value{
data.frame
}
\description{
Import BiSSE log files to get estimated rates
}
