% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mcmc.R
\name{read_mcmc_logs}
\alias{read_mcmc_logs}
\title{Read MCMC logs}
\usage{
read_mcmc_logs(path, pattern, burnin = 0.1, different_lengths = FALSE)
}
\arguments{
\item{path}{directory in which log files are located}

\item{pattern}{string to match for files to include in list}

\item{burnin}{default 0.1}

\item{different_lengths}{if mcmc chains are different lengths in set to avoid running slowly}
}
\value{
logs list of mcmc posteriors
}
\description{
Read mcmc logs into list of posteriors
}
