% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{sample_alive_cells}
\alias{sample_alive_cells}
\title{Sample alive cells}
\usage{
sample_alive_cells(
  alive_cells,
  n,
  diversified_sampling = FALSE,
  weighted_sampling = FALSE,
  edge_weight = 0.5,
  fixed_edge_num = NULL
)
}
\arguments{
\item{alive_cells}{data.frame}

\item{n}{integer}

\item{diversified_sampling}{logical for diversified sampling}

\item{edge_weight}{probability of edge sample for diversified sampling -- parameter for binomial distribution}

\item{fixed_edge_num}{number of edge cells to sample, if not provided number of edge samples are randomly drawn}
}
\value{
data.frame
}
\description{
Randomly sample alive cells.
}
