% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_sample_sim.R
\name{shift_punch}
\alias{shift_punch}
\title{Shift cells}
\usage{
shift_punch(proposed_sampled_cell_index, alive_cells)
}
\arguments{
\item{proposed_sampled_cell_index}{integer, original punch index to shift}

\item{alive_cells}{data.frame of alive cells in simulated tumor}
}
\value{
integer, index cell of new proposed center of punch
}
\description{
Utility function for bulk sampling tumor, will shift punch over if there are not enough cells sampled. This is mostly useful for sampling near edge.
}
