% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateClocks.R
\name{write_state_clocks_xml}
\alias{write_state_clocks_xml}
\title{Write BEAST StateClocks XML}
\usage{
write_state_clocks_xml(
  sim_cells_file,
  sampled_cells_file,
  xml_file,
  template_file,
  rewrite_sampled_cells_csv = FALSE,
  n_samples = 100,
  diversified = FALSE,
  resample = FALSE
)
}
\arguments{
\item{sim_cells_file}{string File for csv of all simulated cells}

\item{xml_file}{string File to write new XML file}

\item{template_file}{string Filename of XML template that includes insert_sequence and insert_trait placeholders}

\item{rewrite_sampled_cells_csv}{logical FALSE. Rewrite sampled cells file with edge/center info}

\item{n_samples}{integer default 100. Number of cells to sample if sampling.}

\item{diversified}{logical default FALSE. If sampling new cells should be diversified where distance is maximized between sampled cells.}

\item{sampled_cell_file}{string File for csv of only sampled cells. If doesn't exist, cells will be sampled and newly sampled cells will be written to this file.}
}
\description{
Insert sequence and trait data from simulated tumor cells into template xml to run BEAST StateClocks
Template must include "insert_sequence" to indicate spot for alignment data and "insert_trait" to indicate location for cell states.
}
