% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{convert_all_cells_to_tree_fast}
\alias{convert_all_cells_to_tree_fast}
\title{Convert all cells to tree fast}
\usage{
convert_all_cells_to_tree_fast(
  all_cells,
  add_all_states = FALSE,
  sampled_cells_indices = NULL,
  branch_unit = "time",
  cell_locations_df_file = NULL
)
}
\arguments{
\item{all_cells}{data.frame of all cells in simulation}

\item{add_all_states}{Default: FALSE. If TRUE, will find all edge/center states at cell death. Takes some time.}

\item{cell_locations_df_file}{Default: Null. CSV file with locations of cells through simulation time.}

\item{sampled_cell_indices}{Default: Null. Vector of sampled cells if making multitype tree with single child nodes}
}
\value{
treedata object with simulated tree
}
\description{
Convert all simulated cells to tree data using faster algorithm
}
