% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_info.R
\name{find_cell_state_at_death}
\alias{find_cell_state_at_death}
\title{Find cell state at death}
\usage{
find_cell_state_at_death(index, all_cells)
}
\arguments{
\item{index}{integer}

\item{all_cells}{data.frame}
}
\value{
integer
}
\description{
Find binary edge/center spatial state at time of cell division or death
}
