% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumortree.R
\name{prune_simulated_tree}
\alias{prune_simulated_tree}
\title{Prune simulated tree}
\usage{
prune_simulated_tree(
  tree,
  sampled_cells_indices,
  add_all_states = FALSE,
  all_cells = NULL
)
}
\arguments{
\item{tree}{treedata object of simulated tree}

\item{add_all_states}{Default: FALSE. If TRUE, will find all edge/center states at cell death. Takes some time.}

\item{all_cells}{data.frame of simulated cells, required for state assignment.}

\item{sampled_cell_indices}{Vector of indices of sampled cells.}
}
\value{
treedata object with simulated tree
}
\description{
Prune simulated tree while preserving tree data. Output tree will not contain single-child nodes, and state-assignment will be at node level.
}
